backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
umap-learn>=0.5
yfinance>=0.2.65
requests>=2.32
pyarrow>=16.1
python-dateutil>=2.9
